refactor tool icons components
