+++ assets/src/components/workbenches/workbench/job/WorkbenchJob.tsx
@@ -3,16 +3,17 @@ import {
   EmptyState,
   Flex,
   SidePanelOpenIcon,
-  Tooltip,
   useSetBreadcrumbs,
 } from '@pluralsh/design-system'
 import { RunStatusChip } from 'components/ai/infra-research/details/InfraResearch'
 import { POLL_INTERVAL } from 'components/cd/ContinuousDeployment'
 import { GqlError } from 'components/utils/Alert'
 import { Confirm } from 'components/utils/Confirm'
+import { MetadataIcons } from 'components/utils/MetadataIcons'
 import { useSimpleToast } from 'components/utils/SimpleToastContext'
 import { StretchedFlex } from 'components/utils/StretchedFlex'
 import { StackedText } from 'components/utils/table/StackedText'
+import { WorkbenchToolIcon } from 'components/workbenches/tools/workbenchToolsUtils'
 import {
   useCancelWorkbenchJobMutation,
   useWorkbenchJobQuery,
@@ -30,12 +31,9 @@ import {
 import styled, { useTheme } from 'styled-components'
 import { formatDateTime } from 'utils/datetime'
 import { isNonNullable } from 'utils/isNonNullable'
-import { WorkbenchToolIcon } from '../../tools/workbenchToolsUtils'
 import { SaveWorkbenchPromptButton } from '../SaveWorkbenchPromptButton'
 import { WorkbenchJobActivities } from './WorkbenchJobActivities'
 import { useWorkbenchJobPanel } from './WorkbenchJobPanel'
-
-const MAX_VISIBLE_JOB_TOOLS = 3
 
 export function WorkbenchJob() {
   const theme = useTheme()
@@ -86,12 +84,6 @@ export function WorkbenchJob() {
     job?.status === WorkbenchJobStatus.Pending ||
     job?.status === WorkbenchJobStatus.Running
   const jobTools = job?.workbench?.tools?.filter(isNonNullable) ?? []
-  const visibleJobTools = jobTools.slice(0, MAX_VISIBLE_JOB_TOOLS)
-  const hiddenJobTools = jobTools.slice(MAX_VISIBLE_JOB_TOOLS)
-  const hiddenJobToolsLabel = hiddenJobTools
-    .map((tool) => tool.name)
-    .filter(Boolean)
-    .join(', ')
 
   useSetBreadcrumbs(
     useMemo(
@@ -174,32 +166,20 @@ export function WorkbenchJob() {
                     </span>
                   )}
                   {!isEmpty(jobTools) && (
-                    <Flex gap="xsmall">
-                      {visibleJobTools.map((tool) => (
-                        <Tooltip
-                          key={tool.id}
-                          label={tool.name}
-                          placement="bottom"
-                        >
+                    <MetadataIcons
+                      maxVisibleItems={3}
+                      items={jobTools.map((tool) => ({
+                        id: tool.id,
+                        label: tool.name,
+                        icon: (
                           <WorkbenchToolIcon
                             type={tool.tool}
                             provider={tool.cloudConnection?.provider}
                             size={12}
                           />
-                        </Tooltip>
-                      ))}
-                      {!isEmpty(hiddenJobTools) && (
-                        <Tooltip
-                          label={
-                            hiddenJobToolsLabel ||
-                            `${hiddenJobTools.length} more`
-                          }
-                          placement="bottom"
-                        >
-                          <span>+{hiddenJobTools.length}</span>
-                        </Tooltip>
-                      )}
-                    </Flex>
+                        ),
+                      }))}
+                    />
                   )}
                 </Flex>
               )
